Change np.int and np.long

+++ torchfm/dataset/movielens.py
@@ -18,11 +18,11 @@ class MovieLens20MDataset(torch.utils.data.Dataset):
 
     def __init__(self, dataset_path, sep=',', engine='c', header='infer'):
         data = pd.read_csv(dataset_path, sep=sep, engine=engine, header=header).to_numpy()[:, :3]
-        self.items = data[:, :2].astype(np.int) - 1  # -1 because ID begins from 1
+        self.items = data[:, :2].astype(np.int32) - 1  # -1 because ID begins from 1
         self.targets = self.__preprocess_target(data[:, 2]).astype(np.float32)
         self.field_dims = np.max(self.items, axis=0) + 1
-        self.user_field_idx = np.array((0, ), dtype=np.long)
-        self.item_field_idx = np.array((1,), dtype=np.long)
+        self.user_field_idx = np.array((0, ), dtype=np.int64)
+        self.item_field_idx = np.array((1,), dtype=np.int64)
 
     def __len__(self):
         return self.targets.shape[0]
